Cleaned up file organization on data (hard coded arrays for seeds), API routes, DB query file for products and adding more seed logic
